+++ modules/jaspr-code/src/jaspr/scopes_domain.ts
@@ -55,6 +55,7 @@ export class ScopesDomain implements vscode.Disposable {
     this.workspaceSubscriptions = vscode.workspace.onDidChangeWorkspaceFolders(
       (_) => this.registerFolders()
     );
+    this.toolingDaemon.onDidRestart(() => this.registerFolders());
     this.configurationSubscriptions = vscode.workspace.onDidChangeConfiguration(
       (_) => this.updateDiagnostics()
     );
@@ -149,7 +150,10 @@ export class ScopesDomain implements vscode.Disposable {
             let message = `Unsafe import: '${dep.invalidOnServer.uri}' depends on '${dep.invalidOnServer.target}', which is not available on the server.\nTry using a platform-independent library ${messageSuffix}`;
             if (s.uri === "package:jaspr/browser.dart") {
               message = `Unsafe import: '${s.uri}' is not available on the server.\nTry using 'package:jaspr/jaspr.dart' instead ${messageSuffix}`;
-            } else if (s.uri === "package:web/web.dart" || s.uri === "dart:js_interop") {
+            } else if (
+              s.uri === "package:web/web.dart" ||
+              s.uri === "dart:js_interop"
+            ) {
               message = `Unsafe import: '${s.uri}' is not available on the server.\nTry using the 'universal_web' package instead ${messageSuffix}`;
             } else if (dep.invalidOnServer.uri === dep.invalidOnServer.target) {
               message = `Unsafe import: '${dep.invalidOnServer.uri}' is not available on the server.\nTry using a platform-independent library ${messageSuffix}`;