+++ packages/jaspr/tool/generate_html.dart
@@ -5,13 +5,13 @@ void main() {
   var specFile = File('tool/data/html.json');
   var specJson = jsonDecode(specFile.readAsStringSync()) as Map<String, dynamic>;
 
-  var extensionSpecFile = File('../../modules/jaspr-code/src/jaspr/html_spec.ts');
+  var cliSpecFile = File('../jaspr_cli/lib/src/html_spec.dart');
 
-  extensionSpecFile.writeAsString('''
+  cliSpecFile.writeAsString('''
 // GENERATED FILE - DO NOT EDIT
 // Generated from packages/jaspr/tool/generate_html.dart
 
-export const htmlSpec = ${const JsonEncoder.withIndent('  ').convert(specJson)};
+const htmlSpec = ${const JsonEncoder.withIndent('  ').convert(specJson)};
 ''');
 
   var allTags = <String>{};