+++ packages/jaspr_cli/lib/src/html_spec.dart
@@ -1,7 +1,7 @@
 // GENERATED FILE - DO NOT EDIT
 // Generated from packages/jaspr/tool/generate_html.dart
 
-export const htmlSpec = {
+const htmlSpec = {
   "content": {
     "article": {
       "doc": "The &lt;article&gt; HTML element represents a self-contained composition in a document, page, application, or site, which is intended to be independently distributable or reusable (e.g., in syndication). Examples include: a forum post, a magazine or newspaper article, or a blog entry, a product card, a user-submitted comment, an interactive widget or gadget, or any other independent item of content."